+++ lib/src/track/local_audio_track.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter_webrtc/flutter_webrtc.dart' as rtc;
 
-import '../errors.dart';
+import '../exceptions.dart';
 import 'audio_track.dart';
 import 'options.dart';
